Teljesn adatbázis létrejön.

+++ backend/src/main/java/szakdoga/backend/Repository/ReservationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import szakdoga.backend.app.module.Reservation;
 
 @Repository
-public interface ReservationRepository extends CrudRepository<Reservation, Integer> {
+public interface ReservationRepository extends CrudRepository<Reservation, Long> {
 }
